% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDensPlot.R
\name{corDensPlot}
\alias{corDensPlot}
\title{corDensPlot}
\usage{
corDensPlot(expression, activity, file_prefix)
}
\arguments{
\item{expression}{matrix or data.frame of expression values}

\item{activity}{matrix or data.frame of inferred activities}

\item{file_prefix}{string for the file name prefix for the output png file}
}
\value{
Writes the density plot for all regulators' matched activities and expression values Pearson correlations to a png image file.
}
\description{
This function generates a png image file of a density plot of regulators' activity-to-expression Pearson correlations.
}
\examples{
\dontrun{
corDensPlot(reg_expression,reg_activity,"Regulators")
# Generates the file "Regulators_activity_expression_correlation.png" in your working directory.
}
}
