% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalRF.R
\name{finalRF}
\alias{finalRF}
\title{finalRF}
\usage{
finalRF(rfcv_workspace, exp_file, phenotype, prefix, mr_count = 100)
}
\arguments{
\item{rfcv_workspace}{string indicating the path and file name of the .RData
workspace image file from the random forest cross-validation analysis
generated by the rfCrossVal function.}

\item{exp_file}{string indicating the path and file name of the normalized
expression data file.}

\item{phenotype}{string indicating the column name labeling the phenotype of
interest.}

\item{prefix}{string indicating a prefix to be added to all output file
names.}

\item{mr_count}{numeric indicating how many representative master regulators
should be inferred, which is suggested by the random forest cross-validation
plot. Defaults to 100.}
}
\value{
Writes seven output files:
\enumerate{
  \item  A log file recording the standard outputs generated by the function.
  \item  A pdf file for the importance plot of the final MR random forest
  model that indicates the relative importance of each inferred MR.
  \item  A text file with the summary statistics for phenotype prediction
  accuracy based on the final MR random forest model in the training and
  test sets. For binary phenotypes, these results are from
  caret::confusionMatrix, while for continuous phenotypes they are from
  stats:lm linear models.
  \item  A text file listing the inferred representative phenotypic MRs.
  \item  A tab-delimited text file containing the input normalized
  expression data filtered to include only the representative phenotypic MRs.
  \item  A tab-delimited text file containing the activity data (which was
  already present in the input workspace file) filtered to include only the
  representative phenotypic MRs.
  \item  An .RData workspace image file containing all environment variables
  from this analysis. This may be useful if one wants to validate the final
  MR random forest model in an independent dataset by again comparing actual
  phenotype values to those predicted with the MR random forest model.
}
}
\description{
This function carries out the final master regulator (MR) analysis using the
workspace image file from the random forest cross-validation analysis and
compares phenotype prediction based on the final MR random forest model to
the actual phenotypes in both the training and test sets.
}
\examples{
\dontrun{
finalRF("Toy_data_rfcv_workspace.RData","Sim_pheno","Toy_data")
}
}
