% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMRs.R
\name{grabMRs}
\alias{grabMRs}
\title{grabMRs}
\usage{
grabMRs(exp_data, act_data, mr_list, prefix)
}
\arguments{
\item{exp_data}{string, data.frame or matrix. If a string, it must indicate
the path and file name for the expression data file. Otherwise, the given,
pre-existing data.frame or matrix containing the expression data will be
used.}

\item{act_data}{string, data.frame or matrix. If a string, it must indicate
the path and file name for the activity data file. Otherwise, the given,
pre-existing data.frame or matrix containing the activity data will be
used.}

\item{mr_list}{string or vector. If a string, it must indicate
the path and file name for the gene (MR) name list. Otherwise, the given,
pre-existing vector of gene (MR) names will be used.}

\item{prefix}{string indicating a prefix to be added to all output file
names.}
}
\value{
Writes to text file the expression and activity data for the given
list of genes/MRs.
}
\description{
This function will grab the expression and activity data for anyn list of
genes, though it is implemented here for the purpose of grabbing inferred
phenotype MRs.
}
\examples{
\dontrun{
# Run with objects already in the workspace.
grabMRs(gene_exp,gene_act,mrs,"Toy_data")

# Run with names of files to be read.
grabMRs("./Toy_data_expressed_genes_QNorm_INT_expression.txt",
"Toy_data_regulators_activities.txt",
"Toy_data_representative_Sim_pheno_MRs_list.txt",
"Toy_data")
}
}
