% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancePlot.R
\name{importancePlot}
\alias{importancePlot}
\title{importancePlot}
\usage{
importancePlot(rndFor, phenotype, mr_count, prefix)
}
\arguments{
\item{rndFor}{random forest model object generated by the randomForest
function from the randomForest package.}

\item{phenotype}{string indicating the column name of the phenotype of
interest.}

\item{mr_count}{numeric indicating how many representative master regulators
are being inferred by this analysis.}

\item{prefix}{string indicating a prefix to be added to the output file
name.}
}
\value{
Writes the importance plot from the final random forest model with
the representative phenotypic master regulators to a pdf file.
}
\description{
This function writes a pdf file of the importance plot for the final random
forest model with the representative phenotypic master regulators.
}
\examples{
\dontrun{
importancePlot(final_rndFor,"Sim_pheno",100,"Toy_data")
}
}
