% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixQTL.R
\name{matrixQTL}
\alias{matrixQTL}
\title{matrixQTL}
\usage{
matrixQTL(
  snp_dose,
  snp_map,
  gene_data,
  gene_map,
  covars = NULL,
  prefix,
  cisP = 1,
  transP = 0.05
)
}
\arguments{
\item{snp_dose}{string indicating the path and file name for the SNP dosage
file.}

\item{snp_map}{string indicating the path and file name for the SNP map file.}

\item{gene_data}{string indicating the path and file name for the gene
exression/activities file.}

\item{gene_map}{string indicating the path and file name for the gene map
file.}

\item{covars}{string indicating the path and file name for the covariables
file. If you do not wish to include covariables in the QTL analysis (not
typically recommended), do not assign anything to this argument. Defaults to
NULL.}

\item{prefix}{string indicating a prefix to be added to all output file
names.}

\item{cisP}{numeric indicating the P-value threshold for cis-QTLs to be
written to the output file. Defaults to 1 (all cis-QTL results are output).}

\item{transP}{numeric indicating the P-value threshold for trans-QTLs to be
written to the output file. Defaults to 0.05 (only nominally significant
QTLs are output).}
}
\value{
Writes four output files:
\enumerate{
  \item  A log file recording the standard outputs generated by the function.
  \item  A pdf file for the QQ-plot of all tested cis and trans-QTLs.
  \item  A text file with the cis-QTL summary statistics that includes only
  those that passed the cis P threshold set by the cisP argument.
  \item  A text file with the trans-QTL summary statistics that includes
  only those that passed the trans P threshold set by the transP argument.
}
}
\description{
This is the wrapper function for the eQTL/aQTL analyses using the MatrixEQTL
package, which was authored and is maintained by Andrey A Shabalin
and is under an LGPL-3 license. For MatrixEQTL troubleshooting and bug
reports, visit https://github.com/andreyshabalin/MatrixEQTL.
}
\examples{
\dontrun{
matrixQTL(
"GEUVADIS_filtered_samples_WHR_significant_bi-allelic_SNPs.dosage",
"GEUVADIS_WHR_significant_bi-allelic_SNPs_locations_in_GRCh37.map",
"Toy_data_regulators_activities.txt",
"GEUVADIS_expressed_genes_locations_in_GRCh37.map",
"GEUVADIS_filtered_samples_select_covars.txt",
prefix="Toy_data")
}
}
