% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRaQTL-package.R
\docType{package}
\name{MRaQTL-package}
\alias{MRaQTL}
\alias{MRaQTL-package}
\title{MRaQTL: Expression regulator activity inference with master regulator and activity QTL analysis}
\description{
This package includes all functions required to apply the master regulator (MR) analysis and activity QTL analyses as first demonstrated in Hoskins et al., 2021 (https://doi.org/10.1371/journal.pcbi.1009563) and then developed into a user-friendly pipeline for STAR Protocols (currently under review). The pipeline is divided into 4 general steps. 1. Tissue-specific gene co-expression network inference with ARACNe (not implemented in R). 2. Expression regulators activity inferences and data preparation (implemented in this package). 3. Phenotypic master regulator (MR) inference (implemented in this package). 4. Expression and activity QTL (eQTL and aQTL) analyses (implemented in this package).
}
\keyword{internal}
