% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoAssoc.R
\name{phenoAssoc}
\alias{phenoAssoc}
\title{phenoAssoc}
\usage{
phenoAssoc(
  activities,
  phenotype,
  binary = FALSE,
  min_regs = 100,
  max_regs = 500,
  Bonf_thresh = 0.05
)
}
\arguments{
\item{activities}{data.frame containing the training set activities}

\item{phenotype}{data.frame containing the training set phenotype data}

\item{binary}{logical indicating whether the phenotype is binary. If FALSE,
the phenotype is assumed to be continuous. Categorical data with two levels
will be treated as binary, but more than two levels will not work. Defaults
to FALSE.}

\item{min_regs}{numeric indicating the minimum number of best associated
regulators to use in the cross-validation random forest analysis. Defaults
to 100.}

\item{max_regs}{numeric indicating the maximum number of best associated
regulators to use in the cross-validation random forest analysis. Defaults
to 500.}

\item{Bonf_thresh}{numeric indicating the Bonferroni threshold for
significance that will determine how many of the best associated regulators
to use in the cross-validation random forest analysis as long as the number
lies within the given min/max. Defaults to 0.05.}
}
\value{
Returns a vector of names for the best associated regulators that
will be used in the cross-validation random forest analysis
}
\description{
This function determines the association between the given phenotype and all
input training set activities and then returns a list of the top
Bonferroni-significant regulators (within a given min/max).
}
\examples{
\dontrun{
phenoAssoc(train_act,train_pheno,prefix="Toy_data")
}
}
