% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepActExp.R
\name{prepActExp}
\alias{prepActExp}
\title{prepActExp}
\usage{
prepActExp(gene_exp, covariables, network, prefix)
}
\arguments{
\item{gene_exp}{string for the file name of the expression data file (should
be appropriately normalized and pre-filtered for expression).}

\item{covariables}{string for the file name of the covariables data file.}

\item{network}{string for the file name of an ARACNe network file. IMPORTANT:
this file cannot have a header!!!}

\item{prefix}{string indicating a prefix to be added to all output file
names.}
}
\value{
Writes six output files:
\enumerate{
  \item  A log file recording the standard outputs generated by the function:
  \item  A tab-delimited file for the interactome/regulon generated by the
  aracne2regulon() function from the input ARACNe network and expression
  data.
  \item  A png image file of a density plot for all regulators' matched
  activities and expression values Pearson correlations.
  \item  A tab-delimited file containing the quantile normalized and inverse
  normal transformed expression values for all input genes.
  \item  A tab-delimited file containing the quantile normalized and inverse
  normal transformed expression values for only expression regulator genes
  for which activities were successfully inferred.
  \item  A tab-delimited file containing the activities inferred for
  expression regulator genes by VIPER.
}
}
\description{
This is the wrapper function that infers regulator activities with the VIPER
package, upper quantile normalizes and inverse normal transforms the
expression data, and ensures the covariables, activity and expression data
are properly formatted for downstream master regulator and eQTL/aQTL
analyses.
}
\examples{
\dontrun{
prepActExp(
"GEUVADIS_filtered_samples_expressed_genes_logRPKMs_with_symbols.txt",
"GEUVADIS_filtere_samples_select_covars.txt",
"GEUVADIS_100boots_ARACNe_network_no_header.txt",
"Toy_data")
}
}
