% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfCrossVal.R
\name{rfCrossVal}
\alias{rfCrossVal}
\title{rfCrossVal}
\usage{
rfCrossVal(
  act_file,
  pheno_file,
  phenotype,
  seed = 123,
  train_pct = 0.7,
  min_regs = 100,
  max_regs = 500,
  Bonf_thresh = 0.05,
  replicates = 12,
  prefix
)
}
\arguments{
\item{act_file}{string indicating the path and file name for the regulator
activities previously inferred by VIPER using the prepActExp function in
this package.}

\item{pheno_file}{string indicating the path and file name for the phenotype
data with the same sample in the same order as the regulator activities file.}

\item{phenotype}{string indicating the column name for the phenotype of
interest in the phenotype file.}

\item{seed}{numeric indicating the RNG seed. Defaults to 123}

\item{train_pct}{numeric indicating the proportion of samples to use in the
training set, while the remaining samples will go to the test set. Defaults
to 0.7.}

\item{min_regs}{numeric indicating the minimum number of regulators to use
in the random forest cross-validation testing for feature counts. This
number of regulators will be used in rfcv if the number of regulators that
associate with the phenotype at a Bonferroni-adjusted P less than the given
threshold is less than min_regs. Defaults to 100.}

\item{max_regs}{numeric indicating the maximum number of regulators to use
in the random forest cross-validation testing for feature counts. This
number of regulators will be used in rfcv if the number of regulators that
associate with the phenotype at a Bonferroni-adjusted P less than the given
threshold is greater than max_regs. Defaults to 500.}

\item{Bonf_thresh}{numeric indicating the significance threshold for
Bonferroni-adjusted P used to identify regulators best associated with the
phenotype of interest. Defaults to 0.05.}

\item{replicates}{numeric indicating the number of independent random forest
cross-validation analyses to run. Higher numbers will tend to generate
cleaner plots (up to a point), but will also add to the analysis time.
Defaults to 12.}

\item{prefix}{string indicating a prefix to be added to all output file
names.}
}
\value{
This will output a png image file of the random forest cross-
validation plot and a .RData workspace image file that can be used in the
final random forest modeling used to identify representative phenotypic
master regulators (MRs).
}
\description{
This is the main function for the random forest cross-validation analysis
wherein top phenotype-associated regulators are used in random forest models
over a range of feature counts to predict the phenotype of interest. By
inspecting the output plot, one may gain a reasonable estimation for the
number of features (in this case regulators) required to minimize prediction
error.
}
\examples{
\dontrun{
rfCrossVal("Toy_data_regulators_activities.txt",
"GEUVADIS_simulated_phenotype.txt",
"Sim_pheno",
prefix="Toy_data")
}
}
