% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvWrapper.R
\name{rfcv.wrapper}
\alias{rfcv.wrapper}
\title{rfcv.wrapper}
\usage{
rfcv.wrapper(train_x, train_y, seed)
}
\arguments{
\item{train_x}{data.frame containing the activities for the top phenotype
associated regulators for the training set.}

\item{train_y}{data.frame containing the phenotype data for the training set.}

\item{seed}{numeric indicating the RNG seed.}
}
\value{
Returns a list with the following components:
list(n.var=n.var,error.cv=error.cv,predicted=cv.pred)
}
\description{
This is a wrapper function for calling the rfcv function in the randomForest
package with a given seed. This used for multi-threading the rfcv analysis.
}
\examples{
\dontrun{
cv_rndFor <- rfcv.wrapper(train_zact,train_pheno,seed=1)
}
}
