% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvPlot.R
\name{rfcvPlot}
\alias{rfcvPlot}
\title{rfcvPlot}
\usage{
rfcvPlot(rfcv_list, phenotype, prefix)
}
\arguments{
\item{rfcv_list}{list of length 2 or more of rfcv outputs.}

\item{phenotype}{string indicating phenotype of interest, which is just used
in plot title.}

\item{prefix}{string indicating the output file prefix.}
}
\value{
Writes a png file of the random forest cross-validation plot that
can indicate how many regulators are required to minimize prediction error.
}
\description{
This function determines summary statistics for a list of rfcv objects
across all different feature counts, and from them, creates a png image of
the random forest cross-validation plot showing prediction error as a
function of feature counts.
}
\examples{
\dontrun{
rfcvPlot <- rfcvPlot(cv_rndFor,"Sim_pheno","Toy_data")
}
}
