% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTestPlot.R
\name{trainTestPlot}
\alias{trainTestPlot}
\title{trainTestPlot}
\usage{
trainTestPlot(train_pheno, test_pheno, phenotype, binary = FALSE, prefix)
}
\arguments{
\item{train_pheno}{data.frame containing the training set phenotype data.}

\item{test_pheno}{data.frame containing the test set phenotype data.}

\item{phenotype}{string indicating the column name for the phenotype of
interest.}

\item{binary}{logical indicating whether the phenotype is binary. If FALSE,
the phenotype is assumed to be continuous. Categorical data with two levels
will be treated as binary, but more than two levels will not work. Defaults
to FALSE.}

\item{prefix}{string indicating the output file prefix that will be added to
the file name along with the column name for the phenotype of interest.}
}
\value{
Writes the phenotype distributions plot to png file.
}
\description{
This function generates a plot that compares the phenotype distributions of
the training and test sets to confirm they are comparable. A barplot is used
for binary phenotypes while a density plot is used for continuous phenotypes.
}
\examples{
\dontrun{
trainTestPlot(train_phenos,test_phenos,"Sim_pheno",prefix="Toy_data")
}
}
