% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscale.R
\name{zscale.genes}
\alias{zscale.genes}
\title{zscale.genes}
\usage{
zscale.genes(genes)
}
\arguments{
\item{genes}{matrix or data.frame containing gene metrics with genes in rows and samples in columns. Gene and sample identifiers should be in rownames and colnames, respectively.}
}
\value{
Returns an expression set object (from DESeq2) with the gene metrics Z-scaled across samples.
}
\description{
This function applies Z-scaling for all genes across samples very rapidly using multi-threading when possible.
}
\examples{
\dontrun{
z_genes <- zscale.genes(genes)
}
}
