% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscale.R
\name{zscale.genes}
\alias{zscale.genes}
\title{zscale.genes}
\usage{
zscale.genes(genes, exprSet = F)
}
\arguments{
\item{genes}{matrix or data.frame containing gene metrics with genes in rows
and samples in columns. Gene and sample identifiers should be in rownames
and colnames, respectively.}

\item{exprSet}{logical indicating whether the object returned should be an
ExpressionSet object (when TRUE), or a data.frame (when FALSE). Defaults to
FALSE.}
}
\value{
Returns either an ExpressionSet object (when exprSet=TRUE) or a
data.frame (when exprSet=FALSE) with the gene metrics Z-scaled across
samples.
}
\description{
This function applies Z-scaling for all genes across samples very rapidly
using multi-threading when possible.
}
\examples{
\dontrun{
z_genes <- zscale.genes(genes)
}
}
